+++ app/Http/Controllers/Backend/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Backend;
 
 use App\Http\Controllers\Controller;
 use App\Models\User;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
@@ -135,6 +136,10 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $user = User::find($id);
+        if(! Gate::allows('delete-user',$user)){
+            abort(403);
+        }
         User::destroy($id);
         return redirect()->route('backend.users.index');
     }
